Introduce unauthorized endpoint
It is a simple GET method returning 401 HTTP code with a body saying that either
the SSO token cookie was not found or the token is no more valid.

This is used by the Policy Agent to return "Unauthorized" in the above cases
(previously it would redirect to the OpenAM login page, which is not the desired
behavior since the protected resources are not accessed by human beings).